# Stop using LiniarFifo in BufferedWriter