@daurnimator 2022-01-22
An important test is to write more than the buffer fits (autoflush), then keep going.

I suggest that any good optimizations can be made in `LinearFifo`