@Hejsil 2022-01-22
> I suggest that any good optimizations can be made in `LinearFifo`

I have my doubts we can get the same kinds of performance by optimizing LiniarFifo. As I said, this contruct is more general than what BufferedWriter needs, so it just has to do more work because of that. I'll see if there are some things that can be done though.