@Hejsil 2022-01-22
Also, it's not like using this abstraction made the code simpler or easier to read. What do we actually gain from using it here?