@Hejsil 2022-01-22
> An important test is to write more than the buffer fits (autoflush), then keep going.

Also, this benchmark writes a total of `888840000` bytes to stdout, way more than what fits in the buffer. Autoflushing is definitly being tested here.