@zseri 2022-01-23
could you fix the typo `LiniarFifo` (vs `LinearFifo`) in the commit message?